refactor: redisUtil 인터페이스, 구현체 분리
